rasp/requirements.txt
# Core ML/Vision
opencv-python==4.8.1.78
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.195

# Model Hub
huggingface-hub==0.19.4

# Data Processing
numpy==1.24.3
Pillow==10.0.0
scipy==1.11.4

# Communication
paho-mqtt==1.6.1
requests==2.31.0
pyserial==3.5

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1
